+++ lib/features/todo/domain/bloc/todo_list_state.dart
@@ -13,13 +13,6 @@ class TodoLoaded extends TodoState {
       : todos = (todos..sort((a, b) => a.createdAt.compareTo(b.createdAt)));
 }
 
-class TodoOperationBeingPerformed extends TodoState {
-  final List<Todo> todos;
-  final Todo todoToBeMutated;
-  TodoOperationBeingPerformed(List<Todo> todos, this.todoToBeMutated)
-      : todos = (todos..sort((a, b) => a.createdAt.compareTo(b.createdAt)));
-}
-
 class TodoError extends TodoState {
   final String message;
