+++ lib/features/todo/presentation/common/cubit/todo_operation_notifier.dart
@@ -1,0 +1,6 @@
+import 'package:todo_list/features/todo/domain/entity/todo.dart';
+
+abstract class OperationStatusNotifier {
+  void startOperation(Todo todo);
+  void endOperation();
+}
