+++ lib/features/todo/presentation/todo_single/cubit/todo_single_cubit.dart
@@ -1,5 +1,5 @@
 import 'package:bloc/bloc.dart';
-import 'package:todo_list/config/logging/logger.dart';
+import 'package:todo_list/config/logger/logger.dart';
 import 'package:todo_list/features/todo/domain/entity/todo.dart';
 import 'package:todo_list/features/todo/presentation/common/todo_action.dart';
