+++ pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   flutter_bloc: ^8.1.6
   flutter_localizations:
     sdk: flutter
+  flutter_staggered_animations: ^1.1.1
   flutter_svg: ^2.0.10+1
   get_it: ^7.7.0
   intl: ^0.19.0
